+++ src/components/map/airports/MapAirportsList.vue
@@ -194,6 +194,13 @@ function handleMapClick(e: MapBrowserEvent<any>) {
     if (isMobileOrTablet.value) handlePointerMove(e);
 }
 
+watch(() => String(store.mapSettings.navigraphLayers?.disable) + String(store.mapSettings.navigraphLayers?.gatesFallback), () => {
+    originalAirportsData.value = [];
+    airportsData.value = [];
+
+    setVisibleAirports();
+});
+
 watch(map, val => {
     if (!val || vectorLayer) return;
 
@@ -650,7 +657,12 @@ async function setVisibleAirports() {
             if (!navigraphAirports.every(x => originalAirportsData.value.some(y => y.airport === x.vatsimAirport.icao))) {
                 originalAirportsData.value = [
                     ...originalAirportsData.value,
-                    ...(await Promise.all(navigraphAirports.filter(x => !originalAirportsData.value.some(y => y.airport === x.vatsimAirport.icao)).map(x => $fetch<NavigraphAirportData>(`/api/data/navigraph/airport/${ x.vatsimAirport.icao }?v=${ store.version }&isLayout=${ (store.user?.hasCharts && store.user?.hasFms) ? '1' : '0' }`)))).flatMap(x => x ?? []),
+                    ...(await Promise.all(
+                        navigraphAirports.filter(x => !originalAirportsData
+                            .value
+                            .some(y => y.airport === x.vatsimAirport.icao))
+                            .map(x => $fetch<NavigraphAirportData>(`/api/data/navigraph/airport/${ x.vatsimAirport.icao }?v=${ store.version }&layout=${ (store.user?.hasCharts && store.user?.hasFms && !store.mapSettings.navigraphLayers?.disable) ? '1' : '0' }&originalData=${ store.mapSettings.navigraphLayers?.gatesFallback ? '1' : '0' }`)),
+                    )).flatMap(x => x ?? []),
                 ];
             }